Module lpc55::bootloader::command

Structs

HidHeader

Version

Enums

Command

CommandTag

DataPhase

Signifies which of the three cases of the protocol is used.

FlashReadMargin

Key

numbers can be found in UM, Chap 7.3.2 “Key descriptions”

KeystoreOperation

This is the interface definition to a somewhat limited bootloader API, that then operates on the actual PFR data and PUF periphreal.

ReportId

Response

ResponseTag

Constants

KEYSTORE_KEY_NAMES