Trait lpc55::bootloader::UuidSelectable

pub trait UuidSelectable: Sized {
    fn try_uuid(&mut self) -> Result<Uuid>;

    fn list() -> Vec<Self>;

    fn having(uuid: Uuid) -> Result<Self> { ... }
}

Required methods

fn try_uuid(&mut self) -> Result<Uuid>

Returns the UUID associated with the thing, if it has a UUID.

Note: This may modify the thing as it might need to communicate with it.

fn list() -> Vec<Self>

Returns all of the available things.

Provided methods

fn having(uuid: Uuid) -> Result<Self>

Returns the thing with the given UUID.

Default implementation; replace for better error messages.